Split router into admin and games (#22)
## Describe Your Changes
Split the routers into administration and games.

## Checklist Before Requesting a Review
- [x] The code runs successfully.

```commandline
2023-10-26 22:08:01 prijatelitree-web-1       | INFO:     Started server process [56]
2023-10-26 22:08:01 prijatelitree-web-1       | INFO:     Waiting for application startup.
2023-10-26 22:08:01 prijatelitree-web-1       | INFO:     Application startup complete.
2023-10-26 22:08:42 prijatelitree-web-1       | INFO:     192.168.65.1:62229 - "GET / HTTP/1.1" 200 OK
```
