+++ prijateli_tree/app/routers/games.py
@@ -1,0 +1,58 @@
+from http import HTTPStatus
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+
+from prijateli_tree.app.database import SessionLocal
+from prijateli_tree.app.schemas import Game, GameCreate, PlayerCreate
+
+
+router = APIRouter()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+@router.post("/")
+def route_create_game(
+    game_data: GameCreate,
+    db: Session = Depends(get_db),
+):
+    return {"status": "success"}
+
+
+@router.post("/game/{game_id}/player/")
+def route_add_player(
+    game_id: int, player_data: PlayerCreate, db: Session = Depends(get_db)
+):
+    return {"status": "success"}
+
+
+@router.get("/{game_id}")
+def route_game_access(game_id: int):
+    game = Game.query().filter_by(id=game_id).one_or_none()
+    if game is None:
+        raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="game not found")
+    return {"game_id": game_id}
+
+
+@router.get("{game_id}/player/{player_id}")
+def route_game_player_access(game_id: int, player_id: int):
+    game = Game.query().filter_by(id=game_id).one_or_none()
+    if game is None:
+        raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="game not found")
+
+    if len([player for player in game.players if player.user_id == player_id]) != 1:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail="player not found in game"
+        )
+
+
+@router.post("{game_id}/player/{player_id}/answer")
+def route_add_answer(game_id: int, player_id: int, player_answer: str):
+    pass
