--- prijateli_tree/app/views/administration.py
@@ -1,2 +1,0 @@
-def stuff():
-    print("things")
